requirements.txt
newrelic>=9.2.0
pyyaml>=6.0.1
jinja2>=3.1.3
click>=8.1.7
requests>=2.31.0
rich>=13.7.1
pydantic>=2.5.3
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
redis>=5.0.1
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
structlog>=23.2.0
prometheus-client>=0.19.0
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpx>=0.26.0
